# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Daytask Documentation",
        "contact": {
            "email": "matinnjt2000@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/project": {
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Create a new project",
                "operationId": "storeProject",
                "requestBody": {
                    "request": "StoreProjectRequest",
                    "description": "Project creation details",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "detail",
                                    "start_date"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "New Project"
                                    },
                                    "detail": {
                                        "type": "string",
                                        "example": "New Project Detail"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-08-01 00:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Project created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Project created!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error creating project",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error creating project!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/project/{project}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get a specific project by its ID",
                "operationId": "showProject",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "The ID of the project to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "50"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Project not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Project not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Update a specific project by its ID",
                "operationId": "updateProject",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "The ID of the project to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "20"
                        }
                    }
                ],
                "requestBody": {
                    "request": "UpdateProjectRequest",
                    "description": "Project update details",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "title",
                                    "detail"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Project"
                                    },
                                    "detail": {
                                        "type": "string",
                                        "example": "Updated Project Detail"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-08-01 00:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Project updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Project updated!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error updating project",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error updating project!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Delete a specific project by its ID",
                "operationId": "deleteProject",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "The ID of the project to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "20"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Project deleted!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error deleting project",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error deleting project!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/assign-task-to-project/{project}": {
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Assign tasks to a specific project by its ID",
                "operationId": "assignTask",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "The ID of the project to assign tasks",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "21"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Task assignment details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tasks"
                                ],
                                "properties": {
                                    "tasks": {
                                        "type": "object",
                                        "example": {
                                            "0": "1",
                                            "1": "2",
                                            "3": "4"
                                        },
                                        "additionalProperties": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tasks assigned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tasks assigned!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error assigning tasks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error assigning tasks!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/unassign-task-from-project/{project}/{task}": {
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Unassign a task from a specific project by its ID",
                "operationId": "unassignTask",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "The ID of the project to unassign the task from",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "21"
                        }
                    },
                    {
                        "name": "task",
                        "in": "path",
                        "description": "The ID of the task to unassign from the project",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task unassigned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Task unassigned!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error unassigning task",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error unassigning task!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ongoing-projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get all ongoing projects for the authenticated user",
                "operationId": "ongoingProjects",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Error retrieving projects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error retrieving projects!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/completed-projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get all completed projects for the authenticated user",
                "operationId": "completedProjects",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Error retrieving projects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error retrieving projects!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}